third_party/blink/web_tests/virtual/document-transition/document-transition/capture-callback-exception-expected.txt

CONSOLE ERROR: Uncaught (in promise) ReferenceError: foo is not defined
CONSOLE ERROR: Uncaught (in promise) AbortError: Transition was skipped
